move PuzzlePiece type into an independent file
