+++ src/types/index.ts
@@ -1,0 +1,4 @@
+ export type PuzzlePiece = {
+    letter: string,
+    status: "absent" | "partial" | "correct" | "empty",
+}
